# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MyCarQR API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/vehicle/{qrCode}:
    get:
      operationId: getPublicVehicle
      tags: [public]
      summary: Get public vehicle info by QR code (no auth)
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public vehicle info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicVehicleInfo"
        "404":
          description: Not found or QR disabled
  /public/vehicle/{qrCode}/sos:
    get:
      operationId: getPublicSos
      tags: [public]
      summary: Get SOS emergency profile for a vehicle's owner (no auth)
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SOS profile info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSosInfo"
        "404":
          description: Not found or SOS disabled
  /public/vehicle/{qrCode}/accident:
    post:
      operationId: submitAccidentReport
      tags: [public]
      summary: Submit an accident witness report (no auth)
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAccidentBody"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccidentReport"
  /public/vehicle/{qrCode}/lost-item:
    post:
      operationId: submitLostItem
      tags: [public]
      summary: Submit a lost item/key found report (no auth)
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitLostItemBody"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
  /public/vehicle/{qrCode}/alert:
    post:
      operationId: sendPublicAlert
      tags: [public]
      summary: Send a scan alert to the vehicle owner (no auth)
      parameters:
        - name: qrCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAlertBody"
      responses:
        "201":
          description: Alert sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanAlert"
...
  /sticker-orders:
    post:
      operationId: createStickerOrder
      tags: [stickerOrders]
      summary: Place a physical sticker order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStickerOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StickerOrder"
    get:
      operationId: getMyStickerOrders
      tags: [stickerOrders]
      summary: Get my sticker orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/StickerOrder"
                required: [orders]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ScanAlert:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        alertType:
          type: string
        message:
          type: string
        scannerLocation:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - vehicleId
        - alertType
        - isRead
        - createdAt
    SendAlertBody:
      type: object
      properties:
        alertType:
          type: string
        message:
          type: string
        scannerLocation:
          type: string
      required:
        - alertType
...
    PublicVehicleInfo:
      type: object
      properties:
        vehicleNumber:
          type: string
        vehicleType:
          type: string
        brand:
          type: string
        model:
          type: string
        color:
          type: string
        ownerName:
          type: string
        privacyMode:
          type: boolean
        primaryContact:
          type: string
        whatsappNumber:
          type: string
        preferredContactMethod:
          type: string
        qrCode:
          type: string
      required:
        - vehicleNumber
        - vehicleType
        - privacyMode
        - qrCode
...
    PublicSosInfo:
      type: object
      properties:
        emergencyContactName:
          type: string
        emergencyPhone:
          type: string
        bloodGroup:
          type: string
        medicalNotes:
          type: string
        altContactName:
          type: string
        altContactPhone:
          type: string
      required:
        - emergencyContactName
        - emergencyPhone
        - bloodGroup
        - medicalNotes
        - altContactName
        - altContactPhone
    AccidentReport:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        description:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
        reportedAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - vehicleId
        - description
        - photos
        - reportedAt
        - isRead
    SubmitAccidentBody:
      type: object
      properties:
        description:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
      required:
        - description
    LostItem:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        message:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
        finderContact:
          type: string
        reportedAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - vehicleId
        - message
        - photos
        - reportedAt
        - isRead
    SubmitLostItemBody:
      type: object
      properties:
        message:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
        finderContact:
          type: string
      required:
        - message
...
    StickerOrder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        vehicleId:
          type: integer
          nullable: true
        vehicleQrCode:
          type: string
          nullable: true
        vehicleNumber:
          type: string
          nullable: true
        customerName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        product:
          type: string
        amount:
          type: integer
        stickerStyle:
          type: string
          nullable: true
        screenshotBase64:
          type: string
          nullable: true
        paymentStatus:
          type: string
        orderStatus:
          type: string
        trackingNumber:
          type: string
          nullable: true
        adminNote:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - customerName
        - phone
        - addressLine1
        - city
        - state
        - pincode
        - product
        - amount
        - paymentStatus
        - orderStatus
        - createdAt
        - updatedAt
    CreateStickerOrderRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        vehicleQrCode:
          type: string
        vehicleNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        product:
          type: string
        amount:
          type: integer
        stickerStyle:
          type: string
        screenshotBase64:
          type: string
      required:
        - customerName
        - phone
        - addressLine1
        - city
        - state
        - pincode
        - product
        - amount
...