# lib/api-zod/src/generated/types/stickerOrder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyCarQR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StickerOrder {
  id: number;
  userId: string;
  vehicleId?: number | null;
  vehicleQrCode?: string | null;
  vehicleNumber?: string | null;
  customerName: string;
  phone: string;
  addressLine1: string;
  addressLine2?: string | null;
  city: string;
  state: string;
  pincode: string;
  product: string;
  amount: number;
  stickerStyle?: string | null;
  screenshotBase64?: string | null;
  paymentStatus: string;
  orderStatus: string;
  trackingNumber?: string | null;
  adminNote?: string | null;
  createdAt: string;
  updatedAt: string;
}